refactor: rename every utils name
