+++ __tests__/DomainTest/models/BenefitArray/ChristmasDDayTest.js
@@ -1,5 +1,5 @@
 import Order from '../../../../src/models/Order';
-import ModelUtils from '../../../../src/models/modules/ModelUtils';
+import ModelsUtils from '../../../../src/models/modules/ModelsUtils';
 
 import CONSTANT from '../../../../src/constants/CONSTANT';
 
@@ -43,7 +43,7 @@ describe('크리스마스 디데이 할인', () => {
     const modelInfo = mockModelInfo(date, items);
 
     //when
-    const benefit = ModelUtils.filterBenefit('christmasDDay', modelInfo);
+    const benefit = ModelsUtils.filterBenefit('christmasDDay', modelInfo);
 
     //then
     expect(benefit).toEqual({
@@ -61,7 +61,7 @@ describe('크리스마스 디데이 할인', () => {
     const modelInfo = mockModelInfo(date, items);
 
     //when
-    const benefit = ModelUtils.filterBenefit('christmasDDay', modelInfo);
+    const benefit = ModelsUtils.filterBenefit('christmasDDay', modelInfo);
 
     //then
     expect(benefit).toBe(undefined);
